add components folder + add display.js & moved app.js inside
