+++ src/components/display.js
@@ -1,0 +1,15 @@
+import propTypes from 'prop-types';
+
+const Display = ({ Total }) => <h1>{Total}</h1>;
+
+export default Display;
+
+
+Display.defaultProps = {
+  Total: '0',
+}
+
+Display.propTypes = {
+  Total: propTypes.string,
+
+}
